Fix/price byte

+++ test/foundry/ArithmeticPriceBook/Unit.t.sol
@@ -25,11 +25,11 @@ contract ArithmeticPriceBookUnitTest is Test {
     }
 
     function _testPriceToIndex(
-        uint128 price,
+        uint256 price,
         bool roundingUp,
         uint16 expectedIndex
     ) private {
-        (uint16 priceIndex, uint128 correctedPrice) = priceBook.priceToIndex(price, roundingUp);
+        (uint16 priceIndex, uint256 correctedPrice) = priceBook.priceToIndex(price, roundingUp);
         assertEq(priceIndex, expectedIndex);
         assertEq(correctedPrice, priceBook.indexToPrice(expectedIndex));
     }
